+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -58,6 +58,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a mixed RayMarching/RayTracing mode for RTReflections and RTGI.
 - Added path tracing support for stacklit material.
 - Added support for surface gradient based normal blending for decals.
+- Added support for tessellation for all master node in shader graph.
 
 ### Fixed
 - Fixed Intensity Multiplier not affecting realtime global illumination.